AG-8108 - Move segments util to its own file
