+++ packages/ag-charts-community/src/chart/series/cartesian/areaUtil.ts
@@ -1,13 +1,10 @@
 import type { InternalAgColorType, Point } from 'ag-charts-core';
-import type { AgSeriesMarkerStyle, AgSeriesSegment } from 'ag-charts-types';
+import type { AgSeriesMarkerStyle } from 'ag-charts-types';
 
 import type { NodeUpdateState } from '../../../motion/fromToMotion';
-import { BBox } from '../../../scene/bbox';
 import type { SizedPoint } from '../../../scene/point';
 import type { Path } from '../../../scene/shape/path';
 import type { Segment } from '../../../scene/shape/segmentedPath';
-import { findRangeExtent } from '../../../util/number';
-import type { ChartAxisDirection } from '../../chartAxisDirection';
 import type { SeriesNodeStyleContext } from '../series';
 import type { SeriesNodeDatum } from '../seriesTypes';
 import type { CartesianSeriesNodeDataContext, CartesianSeriesNodeDatum } from './cartesianSeries';
@@ -21,7 +18,7 @@ import {
     prepareLinePathPropertyAnimation,
     prepareLinePathStrokeAnimationFns,
 } from './lineUtil';
-import { type Scaling, isScaleValid } from './scaling';
+import { isScaleValid } from './scaling';
 
 export type AreaFillPathDatum = {
     readonly spans: LinePathSpan[];
@@ -172,50 +169,4 @@ export function prepareAreaPathAnimation(newData: AreaSeriesNodeDataContext, old
     const fill = prepareAreaFillAnimationFns(status, fillSpans, fillPhantomSpans, fadeMode);
     const stroke = prepareLinePathStrokeAnimationFns(status, strokeSpans, fadeMode);
     return { status, fill, stroke };
-}
-
-export function calculateSegments(segmentation: AgSeriesSegment, scales: { [key in ChartAxisDirection]?: Scaling }) {
-    const { key } = segmentation;
-
-    const yScale = scales['y'];
-    const xScale = scales['x'];
-
-    if (!yScale || !xScale) {
-        return;
-    }
-
-    const seriesHeight = findRangeExtent(yScale.range);
-    const seriesWidth = findRangeExtent(xScale.range);
-
-    const isXDirection = key === 'x';
-    const scale = isXDirection ? xScale : yScale;
-
-    return calculateStopStart(segmentation, scale).map(({ stop, start, ...style }) => {
-        let x = isXDirection ? scale.convert(start) : 0;
-        let y = isXDirection ? 0 : scale.convert(start);
-        let width = isXDirection ? scale.convert(stop) - x : seriesWidth;
-        let height = isXDirection ? seriesHeight : scale.convert(stop) - y;
-
-        if (width < 0) {
-            x += width;
-            width = -width;
-        }
-        if (height < 0) {
-            y += height;
-            height = -height;
-        }
-
-        return { clipRect: new BBox(x, y, width, height), ...style };
-    });
-}
-
-function calculateStopStart(segmentation: AgSeriesSegment, scale: Scaling) {
-    const domainStart = scale.domain[0];
-    const domainEnd = scale.domain.at(-1);
-
-    return segmentation.segments.map((segment) => {
-        const { start = domainStart, stop = domainEnd, ...styles } = segment;
-
-        return { ...styles, stop, start };
-    });
 }
